migrations from public

+++ functions/ToolScripts/emuDeckESDE.sh
@@ -31,9 +31,10 @@ ESDE_cleanup(){
 # 2.2 migration
 ESDE_migration(){
 
-	if [ -f "${toolsPath}/EmulationStation-DE-x64_SteamDeck.AppImage" ]; then
+	if [ -f "${toolsPath}/EmulationStation-DE-x64_SteamDeck.AppImage" ] && [ ! -L "${toolsPath}/EmulationStation-DE-x64_SteamDeck.AppImage" ]; then
 		mv "${toolsPath}/EmulationStation-DE-x64_SteamDeck.AppImage" "${toolsPath}/EmulationStation-DE.AppImage"
 		sed -i "s|EmulationStation-DE-x64_SteamDeck.AppImage|EmulationStation-DE.AppImage|g" "$toolsPath/launchers/esde/emulationstationde.sh"
+		ln -s  "${toolsPath}/EmulationStation-DE.AppImage" "${toolsPath}/EmulationStation-DE-x64_SteamDeck.AppImage"
 		ESDE_createDesktopShortcut
 	fi
 }
